[UTIL] Add script to set/toggle wallpaper
This script sets the wallpaper based on time. It checks the time and
sets either the dark wallpaper setup or the light wallpaper
setup. This helps in setting the right type of wallpaper for the eyes
on boot automatically. It can also be added to the crontab/systemd
timers to automatically update the wallpaper in an active session.