[THEME] Use custom load-theme function for emacs
A custom `cm/load-theme` function is now used to set the theme for
emacs. On reading the documentation of `load-theme` and `enable-theme`
in Emacs, it seems that the functions don't check if the theme is
already loaded or not. This is presumably to allow reloading quickly
while working on customizing themes. `cm/load-theme` function makes
sure that the theme is loaded only if it has not previously been
loaded. This should, theoretically, provide a performance boost when
setting themes without making any customizations in the themes
itself, for example: switching between/using only standard themes.