[LEMONBAR] Add a fix for lemonbar on startup
The `panel` script was not starting the lemonbar because the script
uses the environment variables PANEL_HEIGHT, PANEL_WM_NAME and
PANEL_FONT. Since bspwmrc is executed before the zshrc.local gets
loaded, these environment variables were not available for the panel
to start. The same issue was also being faced by the `repanel` script
that is used to toggle the bar on the screen. Hence, these variables
are added to `bspwmrc` itself, for now. While this is a hacky way, it
works for now and is good enough for a first version of the
lemonbar. Further work needs to be done to include loading these
variables in the right way and to improve the lemonbar appearance and use.