[BSPWM] Load Xmodmap for custom keys
Due to moving over to Emacs, I'm trying to shift all my WM keys to the
Hyper key which I've set to the Left Control Key. This will hopefully
help also allow me to include the Super key in the Emacs bindings.

Also added a one-shot rule to open the first Emacs window on the 7th
workspace and then the rest wherever it is called.

Add a rule for the Surf browser to be opened in the tiled state since
it opens in a floating window by default.