[BAR] Add labels and different font sizes
I am using the bar on a dual monitor setup and till now was using the
default font size only on both the screens. With this, I'm adding two
font sizes, 11 for the smaller laptop screen and the default for the
monitor as it was working fine on that. I also added labels in the
output of the modules to easily determine the information type in the bar.